game and fix

+++ Rpg_Brawler/player.c
@@ -1,5 +1,6 @@
 #include "player.h"
 #include "defense.h"
+#include "weapon.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -9,7 +10,7 @@ static player_t createPlayer(char *name)
 {
     player_t new_player;
     strcpy(new_player.name, name);
-    new_player.hp = 150;
+    new_player.hp = (int8_t)150;
     new_player.coin_purse = 2000;
     new_player.defense;
